Refactor: Pagination eslint 오류 때문에 이름 변경
