+++ src/components/shared/PaginationBtn/index.tsx
@@ -1,6 +1,6 @@
-import { NextButton, PageList, PrevButton } from '@/components/shared/Pagination/Button';
+import { NextButton, PageList, PrevButton } from '@/components/shared/PaginationBtn/Button';
 
-export default function Pagination({ page, totalPage, handlePageChange }: { handlePageChange: (page: number) => void; page: number; totalPage: number }) {
+export default function PaginationBtn({ page, totalPage, handlePageChange }: { handlePageChange: (page: number) => void; page: number; totalPage: number }) {
   return (
     <div className="flex-center gap-2 py-4">
       <PrevButton page={page} handlePageChange={handlePageChange} />
